requirements.txt
uvicorn==0.33.0
gunicorn==20.1.0
fastapi==0.115.6
pydantic==2.9.2
pydantic-settings==2.8.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.39
asyncpg==0.30.0